# backend/docs/openapi_pending_actions.yaml
openapi: 3.0.0
info:
  title: Pending Actions API (snippet)
  version: 1.0.0
paths:
  /api/me/pending-actions:
    get:
      summary: Get one-time pending actions for current user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Pending invitations and double-checks
          content:
            application/json:
              schema:
                type: object
                properties:
                  invitations:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        planId:
                          type: string
                        fromUserId:
                          type: string
                        message:
                          type: string
                        type:
                          type: string
                        isFirstTime:
                          type: boolean
                  doubleChecks:
                    type: array
                    items:
                      type: object
                      properties:
                        planId:
                          type: string
                        initiatorId:
                          type: string
                        reason:
                          type: string
                        isPending:
                          type: boolean
  /api/invitations/{invitationId}/mark-seen:
    post:
      summary: Mark an invitation as seen (compatible endpoint)
      parameters:
        - name: invitationId
          in: path
          required: true
          schema:
            type: string
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Marked as seen
  /api/betting-plans/{planId}/doublecheck:
    post:
      summary: Creator double-check action for a pending plan
      parameters:
        - name: planId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                action:
                  type: string
                  enum: [confirm, cancel]
                comment:
                  type: string
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Plan confirmed or cancelled
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT